+++ src/estimpy/player/window.py
@@ -153,7 +153,7 @@ class PlayerWindow(QMainWindow):
         super().__init__()
         self._player = player
         self._es_audio = es_audio
-        self._ss_original_file = None
+        self._ss_pre_audio = None
         self._ramp_active = False
         self._ramp_start_time = 0.0
         self._fps = es.cfg['video.export.fps']
@@ -865,12 +865,12 @@ class PlayerWindow(QMainWindow):
 
         # Reload audio with or without stereo stim filtering
         if new_state:
-            self._ss_original_file = self._es_audio.file
+            self._ss_pre_audio = self._es_audio
             with es.utils.Spinner('Applying stereo stim... '):
                 ss_audio = self._es_audio.with_stereo_stim()
         else:
-            ss_audio = es.audio.Audio(file=self._ss_original_file)
-            self._ss_original_file = None
+            ss_audio = self._ss_pre_audio
+            self._ss_pre_audio = None
 
         # Reload into player and rebuild visualization
         self._player.set_audio(es_audio=ss_audio)